requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
dj-database-url==3.0.1
Django==5.2.7
django-celery-beat==2.8.1
django-timezone-field==7.1
django-unfold==0.69.0
djangorestframework==3.16.1
environs==14.3.0
frozenlist==1.8.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
kombu==5.5.4
magic-filter==1.0.12
marshmallow==4.0.1
multidict==6.7.0
nest-asyncio==1.6.0
packaging==25.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.11.10
pydantic_core==2.33.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==4.6.0
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
yarl==1.22.0
